refactor: 기존 UI 폴더 내 파일 정리
