+++ lib/presentation/map/components/place_simple_view.dart
