+++ lib/presentation/travel_detail/travel_detail_view_model.dart
