--- lib/utils/stack.dart
@@ -1,15 +1,0 @@
-class StackCollection<E> {
-  final _list = <E>[];
-
-  void push(E value) => _list.add(value);
-
-  E pop() => _list.removeLast();
-
-  E get peek => _list.last;
-
-  bool get isEmpty => _list.isEmpty;
-  bool get isNotEmpty => _list.isNotEmpty;
-
-  @override
-  String toString() => _list.toString();
-}
